+++ config/routes.rb
@@ -1,4 +1,3 @@
 Messaged::Engine.routes.draw do
-  mount ActionCable.server => '/cable'
   resources :messages
 end