+++ lib/generators/messaged/views/views_generator.rb
@@ -1,17 +1,11 @@
 module Messaged
   module Generators
-    class InstallGenerator < Rails::Generators::Base
+    class ViewsGenerator < Rails::Generators::Base
       def set_source_paths
         @source_paths = [
           File.expand_path('templates', __dir__),
-          File.expand_path('../../../..', __dir__)
+          File.expand_path('../../../..', __dir__),
         ]
-      end
-
-      def copy_initializer_file
-        copy_file \
-          'initializer.rb',
-          'config/initializers/messaged.rb'
       end
 
       def copy_view_templates