+++ lib/messaged/engine.rb
@@ -4,6 +4,10 @@ module Messaged
     require "turbo-rails"
     isolate_namespace Messaged
 
+    config.to_prepare do
+      Messaged.user_class&.send(:include, Messaged::UserExtender)
+    end
+
     # initializer "messaged.assets.precompile" do |app|
     #   app.config.assets.precompile += %w( messaged/application.css )
     # end