+++ test/dummy/config/routes.rb
@@ -1,4 +1,6 @@
 Rails.application.routes.draw do
+  devise_for :users
+  root to: "home#index"
+  mount ActionCable.server => '/cable'
   mount Messaged::Engine => "/messaged"
-  root to: "messaged/messages#index"
 end