+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_10_01_194736) do
+ActiveRecord::Schema[7.0].define(version: 2022_10_02_190631) do
   create_table "action_text_rich_texts", force: :cascade do |t|
     t.string "name", null: false
     t.text "body"
@@ -50,11 +50,43 @@ ActiveRecord::Schema[7.0].define(version: 2022_10_01_194736) do
   end
 
   create_table "messaged_messages", force: :cascade do |t|
+    t.integer "tenant_id"
+    t.integer "user_id"
+    t.integer "messaged_room_id"
     t.text "content"
+    t.string "delivery_target"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["messaged_room_id"], name: "index_messaged_messages_on_messaged_room_id"
+    t.index ["tenant_id"], name: "index_messaged_messages_on_tenant_id"
+    t.index ["user_id"], name: "index_messaged_messages_on_user_id"
+  end
+
+  create_table "messaged_rooms", force: :cascade do |t|
+    t.integer "tenant_id"
+    t.integer "user_id"
+    t.string "title"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["tenant_id"], name: "index_messaged_rooms_on_tenant_id"
+    t.index ["user_id"], name: "index_messaged_rooms_on_user_id"
+  end
+
+  create_table "users", force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.integer "account_id"
+    t.string "reset_password_token"
+    t.datetime "reset_password_sent_at"
+    t.datetime "remember_created_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["account_id"], name: "index_users_on_account_id"
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "messaged_messages", "messaged_rooms"
 end