+++ CHANGELOG.md
@@ -9,3 +9,6 @@
 * Properly integrate turbo-rails to Engine
 * Partially implement Collin's rich_text updates
 * Update README with instructions based on current manual items that need to be completed by the developer utilizing this engine
+* Add Devise and ActsAsTenant to dummy app
+* Move views into generator
+* Add compatibility for `current_user` and `current_tenant`